+++ src/test/kotlin/test/modelchecking/MpsCheckTaskTest.kt
@@ -4,14 +4,18 @@ import de.itemis.mps.gradle.ErrorMessages
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 import org.hamcrest.CoreMatchers.containsString
+import org.hamcrest.CoreMatchers.not
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TemporaryFolder
 import support.extractTestProject
+import java.io.File
 
 class MpsCheckTaskTest {
+    val mpsVersion = "2025.1.3"
+
     @Rule
     @JvmField
     val testProjectDir: TemporaryFolder = TemporaryFolder()
@@ -22,7 +26,7 @@ class MpsCheckTaskTest {
         }
     """.trimIndent()
 
-    private fun buildScriptBoilerplate(mpsVersion: String) = """
+    private fun buildScriptBoilerplate() = """
             import de.itemis.mps.gradle.tasks.MpsCheck
 
             plugins {
@@ -55,7 +59,7 @@ class MpsCheckTaskTest {
         extractTestProject("test-project", mpsTestPrjLocation)
 
         settingsFile.writeText(settingsScriptBoilerplate())
-        buildFile.writeText(buildScriptBoilerplate("2021.3.3") + """
+        buildFile.writeText(buildScriptBoilerplate() + """
             val checkProject by tasks.registering(MpsCheck::class) {
                 mpsHome.set(layout.dir(resolveMps.map { it.destinationDir }))
                 junitFile.set(layout.buildDirectory.file("output.xml"))
@@ -77,7 +81,7 @@ class MpsCheckTaskTest {
         extractTestProject("test-project", mpsTestPrjLocation)
 
         settingsFile.writeText(settingsScriptBoilerplate())
-        buildFile.writeText(buildScriptBoilerplate("2021.3.3") + """
+        buildFile.writeText(buildScriptBoilerplate() + """
             val checkProject by tasks.registering(MpsCheck::class) {
                 mpsHome.set(layout.dir(resolveMps.map { it.destinationDir }))
                 projectLocation.set(file("mps-prj"))
@@ -89,7 +93,58 @@ class MpsCheckTaskTest {
         val result = gradleRunner().withArguments("checkProject").build()
 
         Assert.assertEquals(TaskOutcome.SUCCESS, result.task(":checkProject")?.outcome)
+        assertThat("no duplicate actions", result.output, not(containsString("Cannot add an action twice")))
         Assert.assertTrue("output file must exist", testProjectDir.root.resolve("build/output.xml").isFile)
+    }
+
+    @Test
+    fun noDuplicateActions() {
+        val settingsFile = testProjectDir.newFile("settings.gradle.kts")
+        val buildFile = testProjectDir.newFile("build.gradle.kts")
+        val mpsTestPrjLocation = testProjectDir.newFolder("mps-prj")
+
+        extractTestProject("test-project", mpsTestPrjLocation)
+
+        settingsFile.writeText(settingsScriptBoilerplate())
+        buildFile.writeText(buildScriptBoilerplate() + """
+            val checkProject by tasks.registering(MpsCheck::class) {
+                mpsHome.set(layout.dir(resolveMps.map { it.destinationDir }))
+                projectLocation.set(file("mps-prj"))
+                junitFile.set(layout.buildDirectory.file("output.xml"))
+            }
+        """.trimIndent())
+
+        val result = gradleRunner().withArguments("checkProject").build()
+
+        Assert.assertEquals(TaskOutcome.SUCCESS, result.task(":checkProject")?.outcome)
+        assertThat("no duplicate actions", result.output, not(containsString("Cannot add an action twice")))
+        Assert.assertTrue("output file must exist", testProjectDir.root.resolve("build/output.xml").isFile)
+    }
+
+    @Test
+    fun failsOnErrors() {
+        val settingsFile = testProjectDir.newFile("settings.gradle.kts")
+        val buildFile = testProjectDir.newFile("build.gradle.kts")
+        val mpsTestPrjLocation = testProjectDir.newFolder("mps-prj")
+        val junitFile = testProjectDir.root.resolve("build/output.xml")
+
+        extractTestProject("test-project-with-errors", mpsTestPrjLocation)
+
+        settingsFile.writeText(settingsScriptBoilerplate())
+        buildFile.writeText(buildScriptBoilerplate() + """
+            val checkProject by tasks.registering(MpsCheck::class) {
+                mpsHome.set(layout.dir(resolveMps.map { it.destinationDir }))
+                projectLocation.set(file("mps-prj"))
+                junitFile.set(layout.buildDirectory.file("output.xml"))
+            }
+        """.trimIndent())
+
+        val result = gradleRunner()
+            .withArguments("checkProject")
+            .forwardOutput()
+            .buildAndFail()
+        Assert.assertEquals(TaskOutcome.FAILED, result.task(":checkProject")?.outcome)
+        Assert.assertTrue("JUnit-formatted results file should exist", junitFile.exists())
     }
 
     @Test
@@ -102,7 +157,7 @@ class MpsCheckTaskTest {
         extractTestProject("test-project", mpsTestPrjLocation)
 
         settingsFile.writeText(settingsScriptBoilerplate())
-        buildFile.writeText(buildScriptBoilerplate("2021.3.3") + """
+        buildFile.writeText(buildScriptBoilerplate() + """
             val checkProject by tasks.registering(MpsCheck::class) {
                 mpsHome.set(layout.dir(resolveMps.map { it.destinationDir }))
                 projectLocation.set(file("mps-prj"))