+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.31.0
+
+### Changed
+
+- Minimum version of mps-build-backends raised to 1.26.0.
+
+### Fixed
+
+- `MpsCheck`: avoid duplicate action registration errors with MPS 2025.1 and above.
+
 ## 1.30.2
 
 ### Fixed